Port MonotoneChain up to JTS-1.10, plus minor dox improvement (memory-oriented)
git-svn-id: http://svn.osgeo.org/geos/trunk@2474 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ source/index/chain/MonotoneChain.cpp
@@ -14,7 +14,7 @@
  *
  **********************************************************************
  *
- * Last port: index/chain/MonotoneChain.java rev. 1.13 (JTS-1.7)
+ * Last port: index/chain/MonotoneChain.java rev. 1.15 (JTS-1.10)
  *
  **********************************************************************/
 
@@ -32,7 +32,7 @@ namespace index { // geos.index
 namespace chain { // geos.index.chain
 
 MonotoneChain::MonotoneChain(const geom::CoordinateSequence *newPts,
-                             int nstart, int nend, void* nContext)
+                             size_t nstart, size_t nend, void* nContext)
 : pts(newPts), env(0), context(nContext), start(nstart), end(nend), id(-1)
 {
 }
